requirements.txt
streamlit
gspread
pandas 
google-auth
streamlit-authenticator==0.1.5
streamlit-option-menu==0.4.0
